create new route to delete cards & tests ok

+++ src/services/CardCreateServices.ts
@@ -39,14 +39,12 @@ class CardCreateServices {
     return update;
   }
 
-  async execute_delete({ id, user_id }: { id: number; user_id: number }) {
-    const [get_card] = await this.cardRepository.getCard({ user_id });
+  async execute_delete({ id, user_id }: { id: number[]; user_id: number }) {
+    const get_card = await this.cardRepository.getCard({ user_id });
 
     if (!get_card) throw new AppError("Nenhum card encontrado");
 
-    if (get_card.user_id !== user_id) throw new AppError("Usuário sem permissão");
-
-    const del_card = await this.cardRepository.delete(id);
+    const del_card = await this.cardRepository.delete({ id, user_id });
 
     return { id: del_card };
   }
